Generate connect code
The server and admin page can now communicate to create a new code for clients to connect with.

This ensures that only the ones at the event can send requests.

In the future, I might make the server support several events happening at the same time, but for now this is made to run on a computer that is stationed at the event.