Client connection
Clients can now connect to the server, and recieve the control panel.

The connection is done with a cookie, containing the pin that the server has generated. If the pin provided in this cookie matches the one the server has, then the client is given access.

Now the next step is to integrate this with all API calls made for the client